requirements.txt
streamlit
tensorflow
pillow
numpy
matplotlib
opencv-python
gdown